refactor: remove `internal/pkg` directory
